Remove old canonical
